refactor: 컴포넌트 폴더 메일 파일 이름 변경
index.tsx
