+++ components/src/lib/components/Modal/Modal.stories.tsx
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from "@storybook/react";
 
-import Modal from "./Modal";
+import Modal from ".";
 
 const meta = {
   title: "Modal",
