+++ components/src/lib/components/Modal/index.tsx
@@ -1,6 +1,6 @@
-import ModalOverlay from "../ModalOverlay/ModalOverlay";
+import ModalOverlay from "../ModalOverlay";
 import ModalContent from "../ModalContent";
-import ModalTitle from "../ModalTitle/ModalTitle";
+import ModalTitle from "../ModalTitle";
 import ModalClose from "../ModalClose";
 import ModalTrigger from "../ModalTrigger";
 import ModalRoot from "../ModalRoot";
